requirements.txt
# Core dependencies
numpy>=1.24.0
scikit-learn>=1.3.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
pypdf>=3.17.0

# NLP and NER
spacy>=3.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Database
pymongo>=4.6.0

# Vector database
chromadb>=0.4.0

# LLM integration
openai>=1.3.0

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

